ARC-000: renaming komponents to react-components
- renamed libs folder `komponents` to `react-components`
- renamed component files to be capitalized
