--- libs/komponents/README.md
@@ -1,7 +1,0 @@
-# komponents
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test komponents` to execute the unit tests via [Vitest](https://vitest.dev/).
