--- libs/komponents/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./lib/header";
-export * from "./lib/user-card";
