+++ libs/react-components/.babelrc
