+++ libs/react-components/README.md
@@ -1,0 +1,7 @@
+# react-components
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test react-components` to execute the unit tests via [Vitest](https://vitest.dev/).
