+++ libs/react-components/project.json
@@ -1,7 +1,7 @@
 {
-  "name": "komponents",
+  "name": "react-components",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/komponents/src",
+  "sourceRoot": "libs/react-components/src",
   "projectType": "library",
   "tags": [],
   "targets": {
@@ -9,7 +9,7 @@
       "executor": "@nx/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/komponents/**/*.{ts,tsx,js,jsx}"]
+        "lintFilePatterns": ["libs/react-components/**/*.{ts,tsx,js,jsx}"]
       }
     },
     "build": {
@@ -17,7 +17,7 @@
       "outputs": ["{options.outputPath}"],
       "defaultConfiguration": "production",
       "options": {
-        "outputPath": "dist/libs/komponents"
+        "outputPath": "dist/libs/react-components"
       },
       "configurations": {
         "development": {
@@ -32,7 +32,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/komponents/jest.config.ts",
+        "jestConfig": "libs/react-components/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
