+++ libs/react-components/src/lib/Header.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 
-import Header from "./header";
+import Header from "./Header";
 
 describe("Header", () => {
   it("should render successfully", () => {
