+++ libs/react-components/tsconfig.lib.json
